Use client key authentication instead of usernames and passwords
This commit replaces the current username+password authentication in the
provider with client ID/client key authentication. This is the preferred
way of authenticating applications with AzureStack, and is thus officially
supported in the ms-rest-azure gem via `ApplicationTokenProvider` class.
This change removes the need for a patch with the custom `PasswordTokenProvider`
we relied on so far.

Fixes #1

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>